🔨Refactor: Move Portfolio domain file to new package
