+++ src/main/java/HYLikeLion/gitppo/gitppoProject/controller/PortfolioApiController.java
@@ -10,10 +10,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import HYLikeLion.gitppo.gitppoProject.domain.term.Portfolio;
+import HYLikeLion.gitppo.gitppoProject.domain.portfolio.Portfolio;
 import HYLikeLion.gitppo.gitppoProject.dto.PortfolioDTO;
 import HYLikeLion.gitppo.gitppoProject.dto.StatusEnum;
-import HYLikeLion.gitppo.gitppoProject.dto.TermDTO;
 
 @RestController
 @RequestMapping("/api/portfolio")
